+++ src/X3DCcdOptimizer/Models/AffinityEvent.cs
@@ -13,7 +13,8 @@ public enum AffinityAction
     DriverSet,
     DriverRestored,
     WouldSetDriver,
-    WouldRestoreDriver
+    WouldRestoreDriver,
+    DetectionSkipped
 }
 
 public record AffinityEvent